feat: add service methods

+++ src/resources/boards/board.service.js
@@ -1,7 +1,7 @@
 import * as boardsRepo from './board.memory.repository.js';
 
 export const getAll = () => boardsRepo.getAll();
-export const getById = () => boardsRepo.getById()
-export const create = () => boardsRepo.create()
-export const update = () => boardsRepo.update()
-export const remove = () => boardsRepo.remove()
+export const getById = ({ userId }) => boardsRepo.getById(userId)
+export const create = ({ body }) => boardsRepo.create(body)
+export const update = ({ userId, body }) => boardsRepo.update(userId, body)
+export const remove = ({ userId }) => boardsRepo.remove(userId)
