+++ src/resources/columns/column.service.js
@@ -1,7 +1,7 @@
 import * as columnsRepo from './column.memory.repository.js';
 
 export const getAll = () => columnsRepo.getAll();
-export const getById = () => columnsRepo.getById()
-export const create = () => columnsRepo.create()
-export const update = () => columnsRepo.update()
-export const remove = () => columnsRepo.remove()
+export const getById = ({ userId }) => columnsRepo.getById(userId)
+export const create = ({ body }) => columnsRepo.create(body)
+export const update = ({ userId, body }) => columnsRepo.update(userId, body)
+export const remove = ({ userId }) => columnsRepo.remove(userId)
