+++ src/resources/tasks/task.service.js
@@ -1,7 +1,7 @@
 import * as tasksRepo from './task.memory.repository.js'
 
 export const getAll = () => tasksRepo.getAll();
-export const getById = () => tasksRepo.getById()
-export const create = () => tasksRepo.create()
-export const update = () => tasksRepo.update()
-export const remove = () => tasksRepo.remove()
+export const getById = ({ userId }) => tasksRepo.getById(userId)
+export const create = ({ body }) => tasksRepo.create(body)
+export const update = ({ userId, body }) => tasksRepo.update(userId, body)
+export const remove = ({ userId }) => tasksRepo.remove(userId)
